dashboard_project/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
